Update stack to include EFS
Creating the stack and connecting to the EC2 successfully allows
mounting the EFS. Will need to see if this can be automated as part
of either the UserData script. Getting the EFS DNS may be possible
as a reference, but only after creation. This may require writing
an external script that sits on S3 that can be modified after the
stack has completed setup.