Fix utility methods for new model training
ImageDataGenerator loads images as "rgb" instead of grayscale
(each channel is just a copy of what already exists). This allows
using the pre-trained weights from imagenet to be used. Hopefully
this will increase the accuracy.

Add versioning and saving the weights of a model to be loaded
by tensorflow/serving as part of the train/checkpoint/save function

Test generated data will use the same 32 batch size because larger
batches seem to crash the GPU and overflow the memory. Not a big
deal with the new model as predictions don't take a long time, so
smaller batch size is inconsequential for predictions.